Templatize plot(), visualizer refactor
By templatizing plot() I also had to templatize the Visualizer class.
Some code inside player was simplified, and now we're sending the
`samples` array (an array of floats) instead of the `mixed` array.

+++ src/player.hpp
@@ -181,7 +181,7 @@ public:                                             \
     MAKE_SIGNAL(cleared, void)
     MAKE_SIGNAL(playlist_changed, List)
     MAKE_SIGNAL(file_removed, int)
-    MAKE_SIGNAL(samples_played, std::span<i16>, std::span<i16>)
+    MAKE_SIGNAL(samples_played, std::span<i16>, std::span<f32>)
     MAKE_SIGNAL(channel_volume_changed, int, int)
 
 #undef MAKE_SIGNAL
